requirements.txt
pytest==7.1.2
allure-pytest==2.9.45
requests==2.27.1
selenium==4.1.5
webdriver_manager==3.5.4
pytest-failed-screenshot==1.0.2
locust==2.8.6